+++ src/main/java/racingcar/Application.java
@@ -1,6 +1,10 @@
 package racingcar;
 
+import org.springframework.jdbc.core.JdbcTemplate;
 import racingcar.controller.GameController;
+import racingcar.dao.JdbcCarDao;
+import racingcar.dao.JdbcGameDao;
+import racingcar.service.RacingCarService;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
 
@@ -12,12 +16,8 @@ public class Application {
         return new InputView(new Scanner(System.in), new OutputView());
     }
 
-    private static OutputView createOutputView() {
-        return new OutputView();
-    }
-
     public static void main(String[] args) {
-        GameController gameController = new GameController(createInputView(), createOutputView());
+        GameController gameController = new GameController(createInputView(), new OutputView());
         gameController.run();
     }
 }
