+++ src/main/java/racingcar/domain/game/GameResultOfCar.java
@@ -1,4 +1,4 @@
-package racingcar.domain.result;
+package racingcar.domain.game;
 
 import java.util.Objects;
