+++ src/main/java/racingcar/domain/game/NumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.domain.numbergenerator;
+package racingcar.domain.game;
 
 public interface NumberGenerator {
     int generate();
