+++ src/test/java/racingcar/domain/car/CarFactoryTest.java
@@ -1,11 +1,12 @@
-package racingcar.domain.carfactory;
+package racingcar.domain.car;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import racingcar.domain.cars.Cars;
+import racingcar.domain.car.CarFactory;
+import racingcar.domain.car.Cars;
 
 import java.util.List;
 import java.util.stream.Stream;
