+++ src/test/java/racingcar/domain/car/CarsTest.java
@@ -1,4 +1,4 @@
-package racingcar.domain.cars;
+package racingcar.domain.car;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -10,6 +10,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
+import racingcar.domain.car.Cars;
 import racingcar.exception.CarsException;
 
 import static org.assertj.core.api.Assertions.assertThat;
